+++ src/main/java/org/boot/growup/auth/persist/entity/Seller.java
@@ -5,8 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.boot.growup.auth.model.UserModel;
 import org.boot.growup.common.constant.Role;
-import org.boot.growup.auth.model.User;
 import org.boot.growup.auth.model.dto.request.SellerSignUpRequestDTO;
 import org.boot.growup.common.entity.BaseEntity;
 import org.boot.growup.product.persist.entity.Brand;
@@ -56,8 +56,8 @@ public class Seller extends BaseEntity {
     @Column(name = "role", nullable = false)
     private Role role;
 
-    public User toUserDetails() {
-        return new User(cpEmail, cpPassword, role);
+    public UserModel toUserDetails() {
+        return new UserModel(cpEmail, cpPassword, role);
     }
 
     /* 판매자 회원가입 */
