+++ src/main/java/org/boot/growup/auth/service/EmailService.java
@@ -1,58 +1,21 @@
 package org.boot.growup.auth.service;
 
 import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import lombok.RequiredArgsConstructor;
 import org.boot.growup.common.model.EmailMessageDTO;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.stereotype.Service;
-import org.thymeleaf.context.Context;
-import org.thymeleaf.spring6.SpringTemplateEngine;
 
-import java.util.Random;
+public interface EmailService {
+    /*
+    이메일 보내기
+     */
+    String sendMail(EmailMessageDTO emailMessage) throws MessagingException;
 
-@Service
-@RequiredArgsConstructor
-public class EmailService {
-    private final JavaMailSender javaMailSender;
-    private final SpringTemplateEngine templateEngine;
+    /*
+    인증코드 생성
+     */
+    String createCode();
 
-    public String sendMail(EmailMessageDTO emailMessage) throws MessagingException {
-        String authCode = createCode();
-
-        MimeMessage mimeMessage = javaMailSender.createMimeMessage();
-
-        MimeMessageHelper mimeMessageHelper = new MimeMessageHelper(mimeMessage, false, "UTF-8");
-        mimeMessageHelper.setTo(emailMessage.getTo()); // 메일 수신자
-        mimeMessageHelper.setSubject(emailMessage.getSubject()); // 메일 제목
-        mimeMessageHelper.setText(setContext(authCode), true); // 메일 본문 내용, HTML 여부
-        javaMailSender.send(mimeMessage);
-
-        return authCode;
-    }
-
-    // 인증번호 및 임시 비밀번호 생성 메서드
-    public String createCode() {
-        Random random = new Random();
-        StringBuffer key = new StringBuffer();
-
-        for (int i = 0; i < 8; i++) {
-            int index = random.nextInt(4);
-
-            switch (index) {
-                case 0: key.append((char) ((int) random.nextInt(26) + 97)); break;
-                case 1: key.append((char) ((int) random.nextInt(26) + 65)); break;
-                default: key.append(random.nextInt(9));
-            }
-        }
-        return key.toString();
-    }
-
-    /* thymeleaf를 통한 html 적용 */
-    public String setContext(String code) {
-        Context context = new Context();
-        context.setVariable("code", code);
-        return templateEngine.process("email", context);
-    }
+    /*
+    thymeleaf를 통한 html 적용
+     */
+    String setContext(String code);
 }
