+++ src/main/java/org/boot/growup/auth/service/SellerService.java
@@ -4,9 +4,11 @@ import org.boot.growup.common.model.TokenDTO;
 import org.boot.growup.auth.model.dto.request.SellerSignInRequestDTO;
 import org.boot.growup.auth.model.dto.request.SellerSignUpRequestDTO;
 import org.boot.growup.auth.persist.entity.Seller;
-import org.springframework.stereotype.Service;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-public interface SellerService {
+public interface SellerService extends UserDetailsService {
     /*
     판매자 이메일 회원가입
      */
@@ -21,5 +23,10 @@ public interface SellerService {
     현재 로그인한 판매자 정보 가져오기.
      */
     Seller getCurrentSeller();
+
+    /*
+    Seller UserDetailsService
+     */
+    UserDetails loadUserByUsername(String username) throws UsernameNotFoundException;
 }
