+++ src/main/java/org/boot/growup/auth/service/impl/SellerServiceImpl.java
@@ -2,13 +2,13 @@ package org.boot.growup.auth.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.boot.growup.auth.model.UserModel;
 import org.boot.growup.auth.service.SellerService;
 import org.boot.growup.common.model.BaseException;
 import org.boot.growup.common.constant.Role;
 import org.boot.growup.common.constant.ErrorCode;
 import org.boot.growup.auth.utils.JwtTokenProvider;
 import org.boot.growup.common.model.TokenDTO;
-import org.boot.growup.auth.service.UserService;
 import org.boot.growup.auth.model.dto.request.SellerSignInRequestDTO;
 import org.boot.growup.auth.model.dto.request.SellerSignUpRequestDTO;
 import org.boot.growup.auth.persist.entity.Seller;
@@ -17,10 +17,12 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
+
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -28,7 +30,6 @@ import org.springframework.util.ObjectUtils;
 public class SellerServiceImpl implements SellerService {
     private final PasswordEncoder passwordEncoder;
     private final SellerRepository sellerRepository;
-    private final UserService userService;
     private final JwtTokenProvider jwtTokenProvider;
 
     @Override
@@ -49,7 +50,7 @@ public class SellerServiceImpl implements SellerService {
 
     @Override
     public TokenDTO signIn(SellerSignInRequestDTO request) {
-        UserDetails userDetails = userService.loadUserByUsername(request.getCpEmail());
+        UserDetails userDetails = loadUserByUsername(request.getCpEmail());
 
         if (!checkPassword(request.getCpPassword(), userDetails.getPassword())) { // 비밀번호 비교
             throw new BaseException(ErrorCode.BAD_REQUEST, "비밀번호가 일치하지 않습니다.");
@@ -84,5 +85,18 @@ public class SellerServiceImpl implements SellerService {
         }
 
         throw new BaseException(ErrorCode.ACCESS_DENIED);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Optional<Seller> seller = sellerRepository.findByCpEmail(username);
+
+        if (seller.isPresent()) {
+            UserModel userDetails = seller.get().toUserDetails();
+            log.info("판매자 권한 {}", userDetails.getAuthorities());
+            return userDetails;
+        }
+
+        throw new UsernameNotFoundException("해당하는 유저를 찾을 수 없습니다.");
     }
 }
